move get_hooks
