+++ crates/vm/levm/src/utils.rs
@@ -8,7 +8,6 @@ use crate::{
         BLOB_GAS_PER_BLOB, COLD_ADDRESS_ACCESS_COST, CREATE_BASE_COST, STANDARD_TOKEN_COST,
         TOTAL_COST_FLOOR_PER_TOKEN, WARM_ADDRESS_ACCESS_COST,
     },
-    hooks::hook::Hook,
     opcodes::Opcode,
     precompiles::{
         is_precompile, SIZE_PRECOMPILES_CANCUN, SIZE_PRECOMPILES_PRAGUE,
@@ -18,7 +17,7 @@ use crate::{
     EVMConfig,
 };
 use bytes::Bytes;
-use ethrex_common::types::{Account, Transaction, TxKind};
+use ethrex_common::types::{Account, TxKind};
 use ethrex_common::{
     types::{tx_fields::*, Fork},
     Address, H256, U256,
@@ -31,17 +30,11 @@ use secp256k1::{
     Message,
 };
 use sha3::{Digest, Keccak256};
-use std::{
-    cell::RefCell,
-    collections::{BTreeSet, HashMap, HashSet},
-    rc::Rc,
-};
+use std::collections::{BTreeSet, HashMap, HashSet};
 pub type Storage = HashMap<U256, H256>;
 
 #[cfg(not(feature = "l2"))]
 use crate::hooks::DefaultHook;
-#[cfg(feature = "l2")]
-use {crate::hooks::L2Hook, ethrex_common::types::PrivilegedL2Transaction};
 
 // ================== Address related functions ======================
 /// Converts address (H160) to word (U256)
@@ -663,31 +656,6 @@ impl<'a> VM<'a> {
         };
 
         Ok(())
-    }
-
-    pub fn get_hooks(_tx: &Transaction) -> Vec<Rc<RefCell<dyn Hook + 'static>>> {
-        #[cfg(not(feature = "l2"))]
-        let hooks: Vec<Rc<RefCell<dyn Hook>>> = vec![Rc::new(RefCell::new(DefaultHook))];
-
-        #[cfg(feature = "l2")]
-        let hooks: Vec<Rc<RefCell<dyn Hook>>> = {
-            let recipient = if let Transaction::PrivilegedL2Transaction(PrivilegedL2Transaction {
-                recipient,
-                ..
-            }) = _tx
-            {
-                Some(*recipient)
-            } else {
-                None
-            };
-
-            vec![Rc::new(RefCell::new(L2Hook {
-                recipient,
-                callframe_backup: CallFrameBackup::default(), // It will be modified afterwards.
-            }))]
-        };
-
-        hooks
     }
 
     /// Gets transaction callee, calculating create address if it's a "Create" transaction.
